move test modules to submodule
