+++ tests/internal/import_hooks/test_integration.py
@@ -63,15 +63,15 @@ class ImportHookTestCase(SubprocessTestCase):
             The import hook should run when the module is imported
         """
         module_hook = mock.Mock()
-        import_hooks.register_module_hook("tests.test_module", module_hook)
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module", module_hook)
 
         # Hook should not be called after register
         module_hook.assert_not_called()
 
-        import tests.test_module
+        import tests.internal.import_hooks.test_module
 
         # Module imported so the hook should be called
-        module_hook.assert_called_once_with(tests.test_module)
+        module_hook.assert_called_once_with(tests.internal.import_hooks.test_module)
 
     def test_import_then_register(self):
         """
@@ -79,13 +79,13 @@ class ImportHookTestCase(SubprocessTestCase):
             The import hook should run immediately when registered
         """
         module_hook = mock.Mock()
-        import tests.test_module
+        import tests.internal.import_hooks.test_module
 
         module_hook.assert_not_called()
-        import_hooks.register_module_hook("tests.test_module", module_hook)
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module", module_hook)
 
         # Hook should be called on register
-        module_hook.assert_called_once_with(tests.test_module)
+        module_hook.assert_called_once_with(tests.internal.import_hooks.test_module)
 
     def test_no_double_call(self):
         """
@@ -93,14 +93,14 @@ class ImportHookTestCase(SubprocessTestCase):
             The import hook should only be run once
         """
         module_hook = mock.Mock()
-        import_hooks.register_module_hook("tests.test_module", module_hook)
-        import tests.test_module
-        import tests.test_module  # noqa
-        import tests.test_module  # noqa
-        import tests.test_module  # noqa
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module", module_hook)
+        import tests.internal.import_hooks.test_module
+        import tests.internal.import_hooks.test_module  # noqa
+        import tests.internal.import_hooks.test_module  # noqa
+        import tests.internal.import_hooks.test_module  # noqa
 
         # Hook should be called only once
-        module_hook.assert_called_once_with(tests.test_module)
+        module_hook.assert_called_once_with(tests.internal.import_hooks.test_module)
 
     def test_register_deregister(self):
         """
@@ -108,9 +108,9 @@ class ImportHookTestCase(SubprocessTestCase):
             The import hook should not be called
         """
         module_hook = mock.Mock()
-        import_hooks.register_module_hook("tests.test_module", module_hook)
-        import_hooks.hooks.deregister("tests.test_module", module_hook)
-        import tests.test_module  # noqa
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module", module_hook)
+        import_hooks.hooks.deregister("tests.internal.import_hooks.test_module", module_hook)
+        import tests.internal.import_hooks.test_module  # noqa
 
         # Hook should not be called
         module_hook.assert_not_called()
@@ -124,22 +124,22 @@ class ImportHookTestCase(SubprocessTestCase):
         test_module_hook2 = mock.Mock()
         test_module2_hook = mock.Mock()
 
-        import_hooks.register_module_hook("tests.test_module", test_module_hook)
-        import_hooks.register_module_hook("tests.test_module", test_module_hook2)
-        import_hooks.register_module_hook("tests.test_module2", test_module2_hook)
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module", test_module_hook)
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module", test_module_hook2)
+        import_hooks.register_module_hook("tests.internal.import_hooks.test_module2", test_module2_hook)
 
         test_module_hook.assert_not_called()
         test_module_hook2.assert_not_called()
         test_module2_hook.assert_not_called()
 
-        import tests.test_module2
+        import tests.internal.import_hooks.test_module2
 
         test_module_hook.assert_not_called()
         test_module_hook2.assert_not_called()
-        test_module2_hook.assert_called_once_with(tests.test_module2)
+        test_module2_hook.assert_called_once_with(tests.internal.import_hooks.test_module2)
 
-        import tests.test_module
+        import tests.internal.import_hooks.test_module
 
-        test_module_hook.assert_called_once_with(tests.test_module)
-        test_module_hook2.assert_called_once_with(tests.test_module)
-        test_module2_hook.assert_called_once_with(tests.test_module2)
+        test_module_hook.assert_called_once_with(tests.internal.import_hooks.test_module)
+        test_module_hook2.assert_called_once_with(tests.internal.import_hooks.test_module)
+        test_module2_hook.assert_called_once_with(tests.internal.import_hooks.test_module2)
